requirements.txt
requests
openai-whisper
zhconv
